backend/python/styling_api/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
scikit-learn
requests
